bugfix: only delete files, if they exist

+++ src/de/uni_koblenz/west/cidre/slave/triple_store/loader/impl/GraphChunkLoader.java
@@ -149,8 +149,10 @@ public class GraphChunkLoader implements GraphChunkListener {
 	}
 
 	private void deleteContent(File workingDir) {
-		for (File containedFile : workingDir.listFiles()) {
-			containedFile.delete();
+		if (workingDir.exists()) {
+			for (File containedFile : workingDir.listFiles()) {
+				containedFile.delete();
+			}
 		}
 	}
